#166 ETH Supply Service
-- Move some logic to common
