+++ supply/ethereum/src/test/kotlin/fund/cyber/supply/ethereum/EthereumSupplyBaseTest.kt
@@ -1,10 +1,11 @@
 package fund.cyber.supply.ethereum
 
 import fund.cyber.common.kafka.BaseKafkaIntegrationTestWithStartedKafka
+import fund.cyber.supply.CommonConfiguration
 import org.springframework.test.context.ContextConfiguration
 import org.springframework.test.context.TestPropertySource
 
 
 @TestPropertySource(properties = ["CHAIN_FAMILY:ETHEREUM", "GENESIS_SUPPLY:72009990.50"])
-@ContextConfiguration(classes = [ApplicationConfiguration::class])
+@ContextConfiguration(classes = [ApplicationConfiguration::class, CommonConfiguration::class])
 abstract class EthereumSupplyBaseTest : BaseKafkaIntegrationTestWithStartedKafka()
